Break apart openevsewifi.js
Split openevsewifi.js into the EVSE code (OpenEVSE control and divert), the services (EmonCMS, MQTT, OhmConnect and OCPP) and the application that pulls them all together.

This allows greater flexability when adding new services and possible expansin to support control of multiple EVSE.
