+++ src/web.js
@@ -36,19 +36,19 @@ app.use(bodyParser.urlencoded({ extended: true }));
 app.get("/config", function (req, res) {
   res.header("Cache-Control", "no-cache, private, no-store, must-revalidate, max-stale=0, post-check=0, pre-check=0");
   res.json({
-    firmware: data.info.firmware,
-    protocol: data.info.protocol,
-    espflash: data.info.espflash,
-    version: data.info.version,
-    diodet: data.openevse.diodet ? 0 : 1,
-    gfcit: data.openevse.gfcit ? 0 : 1,
-    groundt: data.openevse.groundt ? 0 : 1,
-    relayt: data.openevse.relayt ? 0 : 1,
-    ventt: data.openevse.ventt ? 0 : 1,
-    tempt: data.openevse.tempt ? 0 : 1,
-    service: data.openevse.service,
-    scale: data.openevse.scale,
-    offset: data.openevse.offset,
+    firmware: data.evse.info.firmware,
+    protocol: data.evse.info.protocol,
+    espflash: data.evse.info.espflash,
+    version: data.evse.info.version,
+    diodet: data.evse.openevse.diodet ? 0 : 1,
+    gfcit: data.evse.openevse.gfcit ? 0 : 1,
+    groundt: data.evse.openevse.groundt ? 0 : 1,
+    relayt: data.evse.openevse.relayt ? 0 : 1,
+    ventt: data.evse.openevse.ventt ? 0 : 1,
+    tempt: data.evse.openevse.tempt ? 0 : 1,
+    service: data.evse.openevse.service,
+    scale: data.evse.openevse.scale,
+    offset: data.evse.openevse.offset,
     ssid: data.config.wifi.ssid,
     pass: data.config.wifi.pass ? DUMMY_PASSWORD : "",
     emoncms_enabled: data.config.emoncms.enabled,
